feat(Quality Gates): Add more docker image static quality gates
